transferred duplicate files in common
