+++ CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/common/data/converter/UserRatingsConverter.kt
@@ -1,0 +1,12 @@
+package com.ploddasha.cinemashiftapp.common.data.converter
+
+import com.ploddasha.cinemashiftapp.common.data.model.film.UserRatingsItemModel
+import com.ploddasha.cinemashiftapp.common.domain.entity.film.UserRatingsItem
+
+class UserRatingsConverter {
+    fun convertUserRatings(model: UserRatingsItemModel): UserRatingsItem =
+        UserRatingsItem(
+            kinopoisk = model.kinopoisk,
+            imdb = model.imdb
+        )
+}
