+++ CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/common/data/model/film/ActorItemModel.kt
@@ -1,9 +1,9 @@
-package com.ploddasha.cinemashiftapp.film.data.model
+package com.ploddasha.cinemashiftapp.common.data.model.film
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class DirectorItemModel(
+data class ActorItemModel(
     val id: String,
     val professions: List<String>,
     val fullName: String
