+++ CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/common/domain/entity/film/CountryItem.kt
@@ -1,4 +1,4 @@
-package com.ploddasha.cinemashiftapp.film.domain.entity
+package com.ploddasha.cinemashiftapp.common.domain.entity.film
 
 data class CountryItem(
     val name: String,
