+++ CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/common/domain/entity/film/UserRatingsItem.kt
@@ -1,4 +1,4 @@
-package com.ploddasha.cinemashiftapp.film.domain.entity
+package com.ploddasha.cinemashiftapp.common.domain.entity.film
 
 data class UserRatingsItem(
     val kinopoisk: String,
