--- CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/film/data/model/CountryItemModel.kt
@@ -1,11 +1,0 @@
-package com.ploddasha.cinemashiftapp.film.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CountryItemModel(
-    val name: String,
-    val code: String,
-    val code2: String,
-    val id: Int
-)
