--- CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/film/data/model/UserRatingsItemModel.kt
@@ -1,9 +1,0 @@
-package com.ploddasha.cinemashiftapp.film.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class UserRatingsItemModel(
-    val kinopoisk: String,
-    val imdb: String
-)
