--- CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/film/domain/entity/ActorItem.kt
@@ -1,8 +1,0 @@
-package com.ploddasha.cinemashiftapp.film.domain.entity
-
-
-data class ActorItem(
-    val id: String,
-    val professions: List<String>,
-    val fullName: String
-)
