+++ CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/poster/data/converter/FilmItemConverter.kt
@@ -1,17 +1,18 @@
 package com.ploddasha.cinemashiftapp.poster.data.converter
 
-import com.ploddasha.cinemashiftapp.poster.data.model.ActorItemModel
-import com.ploddasha.cinemashiftapp.poster.data.model.CountryItemModel
-import com.ploddasha.cinemashiftapp.poster.data.model.DirectorItemModel
+import com.ploddasha.cinemashiftapp.common.data.converter.ActorConverter
+import com.ploddasha.cinemashiftapp.common.data.converter.CountryConverter
+import com.ploddasha.cinemashiftapp.common.data.converter.DirectorConverter
+import com.ploddasha.cinemashiftapp.common.data.converter.UserRatingsConverter
 import com.ploddasha.cinemashiftapp.poster.data.model.FilmItemModel
-import com.ploddasha.cinemashiftapp.poster.data.model.UserRatingsItemModel
-import com.ploddasha.cinemashiftapp.poster.domain.entity.ActorItem
-import com.ploddasha.cinemashiftapp.poster.domain.entity.CountryItem
-import com.ploddasha.cinemashiftapp.poster.domain.entity.DirectorItem
 import com.ploddasha.cinemashiftapp.poster.domain.entity.FilmItem
-import com.ploddasha.cinemashiftapp.poster.domain.entity.UserRatingsItem
 
 class FilmItemConverter {
+
+    private val actorConverter = ActorConverter()
+    private val directorConverter = DirectorConverter()
+    private val userRatingsConverter = UserRatingsConverter()
+    private val countryConverter = CountryConverter()
 
     fun convert(model: FilmItemModel): FilmItem =
         FilmItem(
@@ -20,41 +21,14 @@ class FilmItemConverter {
             originalName = model.originalName,
             description = model.description,
             releaseDate = model.releaseDate,
-            actorItems = model.actors.map { convertActor(it) },
-            directorItems = model.directors.map { convertDirector(it) },
+            actorItems = model.actors.map { actorConverter.convertActor(it) },
+            directorItems = model.directors.map { directorConverter.convertDirector(it) },
             runtime = model.runtime,
             ageRating = model.ageRating,
             genres = model.genres,
-            userRatingsItem = convertUserRatings(model.userRatings),
+            userRatingsItem = userRatingsConverter.convertUserRatings(model.userRatings),
             img = model.img,
-            countryItem = convertCountry(model.country)
+            countryItem = countryConverter.convertCountry(model.country)
         )
 
-    private fun convertActor(model: ActorItemModel): ActorItem =
-        ActorItem(
-            id = model.id,
-            professions = model.professions,
-            fullName = model.fullName
-        )
-
-    private fun convertDirector(model: DirectorItemModel): DirectorItem =
-        DirectorItem(
-            id = model.id,
-            professions = model.professions,
-            fullName = model.fullName
-        )
-
-    private fun convertUserRatings(model: UserRatingsItemModel): UserRatingsItem =
-        UserRatingsItem(
-            kinopoisk = model.kinopoisk,
-            imdb = model.imdb
-        )
-
-    private fun convertCountry(model: CountryItemModel): CountryItem =
-        CountryItem(
-            name = model.name,
-            code = model.code,
-            code2 = model.code2,
-            id = model.id
-        )
 }
