--- CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/poster/domain/entity/CountryItem.kt
@@ -1,8 +1,0 @@
-package com.ploddasha.cinemashiftapp.poster.domain.entity
-
-data class CountryItem(
-    val name: String,
-    val code: String,
-    val code2: String,
-    val id: Int
-)
