--- CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/poster/domain/entity/DirectorItem.kt
@@ -1,7 +1,0 @@
-package com.ploddasha.cinemashiftapp.poster.domain.entity
-
-data class DirectorItem(
-    val id: String,
-    val professions: List<String>,
-    val fullName: String
-)
