--- CinemaShiftApp/app/src/main/java/com/ploddasha/cinemashiftapp/poster/domain/entity/UserRatingsItem.kt
@@ -1,6 +1,0 @@
-package com.ploddasha.cinemashiftapp.poster.domain.entity
-
-data class UserRatingsItem(
-    val kinopoisk: String,
-    val imdb: String
-)
